requirements.txt
python>=3.6
pytorch>=1.0
ray>=0.6.1
